release 1.05

+++ SwingUI/src/main/java/com/hawolt/async/loader/ResourceLoader.java
@@ -2,7 +2,8 @@ package com.hawolt.async.loader;
 
 import com.hawolt.StaticConstant;
 import com.hawolt.async.ExecutorManager;
-import com.hawolt.cryptography.SHA256;
+import com.hawolt.client.cache.ExceptionalSupplier;
+import com.hawolt.cryptography.MD5;
 import com.hawolt.generic.data.Unsafe;
 import com.hawolt.http.layer.IResponse;
 import com.hawolt.http.layer.impl.OkHttpResponse;
@@ -38,26 +39,32 @@ public class ResourceLoader {
     private static final Map<String, List<ResourceConsumer<?, byte[]>>> pending = new HashMap<>();
     private static final LinkedList<Runnable> queue = new LinkedList<>();
     private static final Map<String, byte[]> cache = new HashMap<>();
+    private static final Object lock = new Object();
 
     static {
-        ScheduledExecutorService scheduler = ExecutorManager.getScheduledService("resource-loader");
-        scheduler.execute(() -> {
+        ScheduledExecutorService executor = ExecutorManager.getScheduledService("cache-loader");
+        executor.execute(() -> {
             File directory = ResourceLoader.directory.toFile();
             if (!directory.exists()) return;
             File[] assets = directory.listFiles();
             if (assets == null) return;
             for (File file : assets) {
                 try {
+                    Logger.debug("FROM CACHE {}", file.getName());
                     cache.put(file.getName(), Files.readAllBytes(file.toPath()));
                 } catch (IOException e) {
                     Logger.error("Failed to load file '{}' from local cache", file.getName());
                 }
             }
         });
+        executor.shutdown();
+        ScheduledExecutorService scheduler = ExecutorManager.getScheduledService("resource-loader");
         scheduler.scheduleWithFixedDelay(() -> {
-            if (queue.isEmpty()) return;
-            service.execute(queue.remove(0));
-        }, 0, 20, TimeUnit.MILLISECONDS);
+            synchronized (lock) {
+                if (queue.isEmpty()) return;
+                service.execute(queue.remove(0));
+            }
+        }, 0, 1, TimeUnit.MILLISECONDS);
     }
 
     private static <T> T convert(Object in) {
@@ -65,7 +72,7 @@ public class ResourceLoader {
     }
 
     private static void load(String path, boolean priority, ResourceConsumer<?, byte[]> consumer, Runnable runnable) {
-        String hash = SHA256.hash(path);
+        String hash = MD5.hash(path);
         if (cache.containsKey(hash)) {
             try {
                 consumer.consume(path, Unsafe.cast(consumer.transform(cache.get(hash))));
@@ -77,7 +84,9 @@ public class ResourceLoader {
         } else {
             pending.put(hash, new ArrayList<>());
             pending.get(hash).add(consumer);
-            queue.add(priority ? 0 : queue.size(), runnable);
+            synchronized (lock) {
+                queue.add(priority ? 0 : queue.size(), runnable);
+            }
         }
     }
 
@@ -85,20 +94,35 @@ public class ResourceLoader {
         loadResource(uri, false, consumer);
     }
 
+    public static void loadResource(String uri, ExceptionalSupplier<byte[]> supplier,
+                                    boolean priority, ResourceConsumer<?, byte[]> consumer) {
+        synchronized (lock) {
+            queue.add(priority ? 0 : queue.size(), () -> load(uri, priority, consumer, () -> {
+                try {
+                    handleConsumption(uri, supplier.get());
+                } catch (Exception e) {
+                    handleError(uri, e);
+                }
+            }));
+        }
+    }
+
     public static void loadResource(String uri, boolean priority, ResourceConsumer<?, byte[]> consumer) {
-        load(uri, priority, consumer, () -> {
-            Request request = new Request.Builder()
-                    .url(uri)
-                    .header("User-Agent", StaticConstant.USER_AGENT)
-                    .get()
-                    .build();
-            try {
-                IResponse response = OkHttpResponse.from(request);
-                handleConsumption(uri, response.response());
-            } catch (IOException e) {
-                handleError(uri, e);
-            }
-        });
+        synchronized (lock) {
+            queue.add(priority ? 0 : queue.size(), () -> load(uri, priority, consumer, () -> {
+                Request request = new Request.Builder()
+                        .url(uri)
+                        .header("User-Agent", StaticConstant.USER_AGENT)
+                        .get()
+                        .build();
+                try {
+                    IResponse response = OkHttpResponse.from(request);
+                    handleConsumption(uri, response.response());
+                } catch (IOException e) {
+                    handleError(uri, e);
+                }
+            }));
+        }
     }
 
     public static void loadLocalResource(String name, ResourceConsumer<?, byte[]> consumer) {
@@ -106,13 +130,15 @@ public class ResourceLoader {
     }
 
     public static void loadLocalResource(String name, boolean priority, ResourceConsumer<?, byte[]> consumer) {
-        load(name, priority, consumer, () -> {
-            try (InputStream stream = RunLevel.get(name)) {
-                handleConsumption(name, Core.read(stream).toByteArray());
-            } catch (IOException e) {
-                handleError(name, e);
-            }
-        });
+        synchronized (lock) {
+            queue.add(priority ? 0 : queue.size(), () -> load(name, priority, consumer, () -> {
+                try (InputStream stream = RunLevel.get(name)) {
+                    handleConsumption(name, Core.read(stream).toByteArray());
+                } catch (IOException e) {
+                    handleError(name, e);
+                }
+            }));
+        }
     }
 
     private static void writeToCache(String o, String hash, byte[] b) {
@@ -121,6 +147,7 @@ public class ResourceLoader {
             try {
                 if (!directory.toFile().exists()) Files.createDirectories(directory);
                 Path target = directory.resolve(hash);
+                if (target.toFile().exists()) return;
                 Files.write(
                         target,
                         b,
@@ -136,7 +163,7 @@ public class ResourceLoader {
     }
 
     private static void handleConsumption(String o, byte[] b) {
-        String hash = SHA256.hash(o);
+        String hash = MD5.hash(o);
         writeToCache(o, hash, b);
         if (!pending.containsKey(hash)) Logger.error("attempt to load unknown value '{}' from cache", o);
         List<ResourceConsumer<?, byte[]>> list = new ArrayList<>(pending.get(hash));
